requirements.txt
streamlit
scikit-learn
python-docx
PyPDF2
